feat: add Windows platform support (v0.5.0)
Add Windows compatibility using the existing #[cfg(target_os)] module
switching pattern. Implements all platform-specific subsystems:

- Packet capture via Winsock2 raw sockets with SIO_RCVALL
- Process table using GetExtendedTcpTable/UdpTable + Toolhelp32
- Interface enumeration via GetAdaptersAddresses
- DNS config from adapter DNS server linked lists
- Signal handling via SetConsoleCtrlHandler

Uses windows-sys 0.59 crate for all Win32 API bindings. No external
packet capture libraries (npcap/WinPcap) required.

Verified: cargo check, cargo clippy, and all 278 tests pass on both
native (Linux) and cross-compiled (x86_64-pc-windows-gnu) targets.

+++ src/capture/linux.rs
@@ -103,7 +103,7 @@ impl AfPacketCapture {
         // 3. Bind to interface
         let mut sll: libc::sockaddr_ll = unsafe { std::mem::zeroed() };
         sll.sll_family = libc::AF_PACKET as u16;
-        sll.sll_protocol = (ETH_P_ALL as u16).to_be();
+        sll.sll_protocol = ETH_P_ALL.to_be();
         sll.sll_ifindex = if_index as i32;
 
         let ret = unsafe {
